requirements.txt
matplotlib==3.7.1
numpy==1.24.2
opencv_python==4.8.1.78
Pillow==9.5.0
Pillow==10.1.0
torch==2.0.1
torchvision==0.15.2
